Answer the code review of the typed access provider seam
The client for an open attempt and the unknown LockStatus move to the
base class, so NUKI and iFBS no longer carry the same mapping twice and
the "unknown on every count" literal lives in one place. iFBS refuses an
open answer without an OpenBox_ID instead of naming "undefined" as the
process. getOpenProgress documents that the access point carries the
tenant, every base method has its @throws, and the changelog entry is
short again.
